Added additional flags and created new function
Added a flag for deleting files after they are copied and also two
flags, one for creating a log of all files that are copied, and another
for deleting those files in that log. These flags are NOT yet
implemented.

Also took out a bunch of the code in main and made it a separate
function.