Moving to Python's built-in logging module
Python's built-in logging module removes a lot of the clutter that's
dirtying up the code I have (otherwise I have a ridiculous number of if
statements that check for a logging flag). This commit only changes over
the parts of the code that write to a logfile, not ones that emit to the
console.